pyproject.toml
[project]
name = "question-app"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "asyncmy>=0.2.10",
    "elasticsearch[async]>=8.17,<8.18",
    "fastapi>=0.116.1",
    "jinja2>=3.1.6",
    "numpy>=2.3.1",
    "ollama>=0.5.1",
    "openai>=1.95.1",
    "pydantic-ai-slim>=0.4,<0.5",
    "pydantic-settings>=2.10.1",
    "pylcs>=0.1.1",
    "pytest-asyncio>=1.0.0",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.13,<1.14",
    "rapidfuzz>=3.13.0",
    "sqlalchemy[asyncio]>=2.0.41",
    "uvicorn[standard]>=0.35.0",
]

[project.scripts]
dev = "question_app.app:run_dev"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]

[tool.pyright]
include = ["src/**/*.py"]
typeCheckingMode = "strict"
reportMissingTypeStubs = false

[tool.ruff]
line-length = 120

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [
    "--capture=no",
    "--import-mode=importlib",
]
asyncio_default_fixture_loop_scope = "function"
asyncio_default_test_loop_scope="function"
asyncio_mode = "auto"
